refactor: categorias

+++ apps/api/src/modules/categorias/controller.ts
@@ -1,22 +1,12 @@
 import { eq } from "drizzle-orm"
 import type { Request, Response } from "express"
-import { z } from "zod"
 import { db } from "~/db/db"
 import { categorias } from "~/db/schema"
 import { paramsSchema } from "~/shared/schemas"
-
-const cadastrarCategoriaSchema = z.object({
-  codigo: z.string().toUpperCase(),
-  nome: z.string().toUpperCase(),
-})
-
-const atualizarCategoriaSchema = z.object({
-  codigo: z.string().toUpperCase().optional(),
-  nome: z.string().toUpperCase().optional(),
-})
+import { createCategoriaSchema, updateCategoriaSchema } from "./dtos"
 
 async function cadastrarCategoria(req: Request, res: Response) {
-  const data = cadastrarCategoriaSchema.parse(req.body)
+  const data = createCategoriaSchema.parse(req.body)
 
   await db.insert(categorias).values(data)
 
@@ -27,7 +17,7 @@ async function cadastrarCategoria(req: Request, res: Response) {
 async function listarCategorias(_: Request, res: Response) {
   const categorias = await db.query.categorias.findMany()
 
-  res.status(200).json({ categorias })
+  res.status(200).json(categorias)
   return
 }
 
@@ -38,6 +28,11 @@ async function mostrarCategoria(req: Request, res: Response) {
     where: (categorias, { eq }) => eq(categorias.id, id),
   })
 
+  if (!categoria) {
+    res.status(404).send({ message: "Categoria não encontrada" })
+    return
+  }
+
   res.status(200).json(categoria)
   return
 }
@@ -45,7 +40,16 @@ async function mostrarCategoria(req: Request, res: Response) {
 async function atualizarCategoria(req: Request, res: Response) {
   const { id } = paramsSchema.parse(req.params)
 
-  const data = atualizarCategoriaSchema.parse(req.body)
+  const categoria = await db.query.categorias.findFirst({
+    where: (categorias, { eq }) => eq(categorias.id, id),
+  })
+
+  if (!categoria) {
+    res.status(404).send({ message: "Categoria não encontrada" })
+    return
+  }
+
+  const data = updateCategoriaSchema.parse(req.body)
 
   await db.update(categorias).set(data).where(eq(categorias.id, id))
 
@@ -55,6 +59,15 @@ async function atualizarCategoria(req: Request, res: Response) {
 
 async function deletarCategoria(req: Request, res: Response) {
   const { id } = paramsSchema.parse(req.params)
+
+  const categoria = await db.query.categorias.findFirst({
+    where: (categorias, { eq }) => eq(categorias.id, id),
+  })
+
+  if (!categoria) {
+    res.status(404).send({ message: "Categoria não encontrada" })
+    return
+  }
 
   await db.delete(categorias).where(eq(categorias.id, id))
